refactor(renderer): split the autolink index and its tests out
The scan and its candidate-start index are separate concerns, and the
autolink edge cases outgrew the shared inline test file. Both files were
over the 350-line limit for changed sources.
